+++ packages/palette-antd/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.2.0-dev.2](https://github.com/react-buddy/ide-toolbox/tree/master/packages/palette-antd/compare/@react-buddy/palette-antd@4.2.0-dev.1...@react-buddy/palette-antd@4.2.0-dev.2) (2022-11-11)
+
+**Note:** Version bump only for package @react-buddy/palette-antd
+
+
+
+
+
+# [4.2.0-dev.1](https://github.com/react-buddy/ide-toolbox/tree/master/packages/palette-antd/compare/@react-buddy/palette-antd@4.2.0-dev.0...@react-buddy/palette-antd@4.2.0-dev.1) (2022-11-08)
+
+**Note:** Version bump only for package @react-buddy/palette-antd
+
+
+
+
+
+# [4.2.0-dev.0](https://github.com/react-buddy/ide-toolbox/tree/master/packages/palette-antd/compare/@react-buddy/palette-antd@4.1.3-dev.0...@react-buddy/palette-antd@4.2.0-dev.0) (2022-11-08)
+
+
+### Features
+
+* add embeddable prop to Palette [#43](https://github.com/react-buddy/ide-toolbox/tree/master/packages/palette-antd/issues/43) ([ad49ad5](https://github.com/react-buddy/ide-toolbox/tree/master/packages/palette-antd/commit/ad49ad52a80cf287c28ba5ccb344f237d4cfa417))
+
+
+
+
+
+## [4.1.3-dev.0](https://github.com/react-buddy/ide-toolbox/tree/master/packages/palette-antd/compare/@react-buddy/palette-antd@4.1.2-dev.0...@react-buddy/palette-antd@4.1.3-dev.0) (2022-10-13)
+
+
+### Bug Fixes
+
+* antd palette cosmetic and wording fixes ([3ff2201](https://github.com/react-buddy/ide-toolbox/tree/master/packages/palette-antd/commit/3ff220155102616b3047c8b2584a8219142230ce))
+
+
+
+
+
 ## [4.1.2](https://github.com/react-buddy/ide-toolbox/tree/master/packages/palette-antd/compare/@react-buddy/palette-antd@4.1.2-dev.0...@react-buddy/palette-antd@4.1.2) (2022-08-10)